requirements.txt
diffusers>=0.4.1
Flask>=1.1.2
numpy>=1.21.6
Pillow>=9.2.0
protobuf>=4.21.6
PyQt5>=5.15.7
requests>=2.28.0
torch>=1.12.1
opencv_python>=4.2.0.32
transformers>=4.22.1
